updating deps.json
